enhance the types of sloots textFiles & sutoComplete
